Hash temporary file names.
Current encoding transforms the file path from where the Globalize calls
were extracted into a file name. In deeply nested projects, this can
double the length of the path, causing E_NAMETOOLONG issues.

Before this:
/long/path/to/project/.tmp-globalize-webpack/-long-path-to-project-src-component-foo-index.js
After this:
/long/path/to/project/.tmp-globalize-webpack/5b4c40dd44a2cfde54a3e04cd183aa70edd64c1a.js

This has the downside that it makes the content of the files more opaque, being
harder to debug which content maps to which file. One would have to
```
echo -n `readlink -f src/component/foo/index.js` | shasum
```

This fixes #42 and #16.